Change Quantity data type from int to float in addStockChangesAPI

+++ store/views.py
@@ -814,7 +814,7 @@ def addStockChangesAPI(request):
             data['tel'] = int(data.get("tel",0))
         else:
             data['tel']=0
-        data['Quantity'] = int(data.get("Quantity"))
+        data['Quantity'] = float(data.get("Quantity"))
         data["timestamp"] = int(datetime.now().timestamp() * 1000)
         type = data.get("type")
         stock_id = data.get("stockId")
